requirements.txt
Django==2.1
psycopg2
djangorestframework
requests